Move PoW related methods and properties to A0BlockHeader from AbtractBlockHeader Move IDifficultyCalculator and IChainCfg to modAion Use A0BlockHeader instead of AbstractBlockHeader with PoW functionalities
