+++ modAion/build.xml
@@ -119,6 +119,7 @@
                 <pathelement location="${dir.mod}/modRlp.jar"/>
                 <pathelement location="${dir.mod}/modCrypto.jar"/>
                 <pathelement location="${dir.mod}/modMcf.jar"/>
+                <pathelement location="${dir.lib}/libnsc.jar"/>
                 <pathelement location="${dir.lib}/commons-lang3-3.4.jar"/>
                 <pathelement location="${dir.lib}/commons-collections4-4.0.jar"/>
                 <pathelement location="${dir.lib}/libJson.jar"/>
