+++ modAion/module-info.java
@@ -10,4 +10,6 @@ module aion.zero {
     exports org.aion.zero.db;
     exports org.aion.zero.types;
     exports org.aion.zero.exceptions;
+    exports org.aion.zero.core;
+    exports org.aion.zero.blockchain;
 }
