+++ modAion/src/org/aion/zero/blockchain/IChainCfg.java
@@ -22,11 +22,12 @@
  *
  ******************************************************************************/
 
-package org.aion.mcf.blockchain;
+package org.aion.zero.blockchain;
 
 import org.aion.base.type.IBlock;
 import org.aion.base.type.ITransaction;
-import org.aion.mcf.core.IDifficultyCalculator;
+import org.aion.mcf.blockchain.IBlockConstants;
+import org.aion.zero.core.IDifficultyCalculator;
 import org.aion.mcf.core.IRewardsCalculator;
 import org.aion.mcf.valid.BlockHeaderValidator;
 import org.aion.mcf.valid.ParentBlockHeaderValidator;
