+++ modAionImpl/src/org/aion/zero/impl/blockchain/ChainConfiguration.java
@@ -26,8 +26,8 @@ package org.aion.zero.impl.blockchain;
 import org.aion.base.type.Address;
 import org.aion.equihash.OptimizedEquiValidator;
 import org.aion.mcf.blockchain.IBlockConstants;
-import org.aion.mcf.blockchain.IChainCfg;
-import org.aion.mcf.core.IDifficultyCalculator;
+import org.aion.zero.blockchain.IChainCfg;
+import org.aion.zero.core.IDifficultyCalculator;
 import org.aion.mcf.core.IRewardsCalculator;
 import org.aion.mcf.mine.IMiner;
 import org.aion.mcf.valid.*;
